Move PythonQuickFixVerifier outside of checks-testkit
