+++ python-checks/pom.xml
@@ -57,6 +57,10 @@
         <artifactId>sonar-analyzer-commons</artifactId>
         <scope>test</scope>
     </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-core</artifactId>
+    </dependency>
   </dependencies>
 
 </project>
